Fix terbilang float-to-int conversion + logo from invoice_logo setting
- Fixed terbilang() function: cast to (int) before division and array access
- Removed PHP deprecation warning about float to int conversion
- Logo now properly fetched from invoice_logo_[businessId] (Report Settings > PDF Logo)
- Fallback chain: invoice_logo_[businessId] -> invoice_logo -> company_logo

+++ includes/print-helper.php
@@ -378,7 +378,7 @@ function getPrintCSS() {
  * Convert number to Indonesian words (terbilang)
  */
 function terbilang($nilai) {
-    $nilai = abs($nilai);
+    $nilai = abs((int)$nilai);
     $huruf = ["", "Satu", "Dua", "Tiga", "Empat", "Lima", "Enam", "Tujuh", "Delapan", "Sembilan", "Sepuluh", "Sebelas"];
     $temp = "";
     
@@ -387,21 +387,21 @@ function terbilang($nilai) {
     } else if ($nilai < 20) {
         $temp = terbilang($nilai - 10) . " Belas";
     } else if ($nilai < 100) {
-        $temp = terbilang($nilai / 10) . " Puluh" . terbilang($nilai % 10);
+        $temp = terbilang((int)($nilai / 10)) . " Puluh" . terbilang($nilai % 10);
     } else if ($nilai < 200) {
         $temp = " Seratus" . terbilang($nilai - 100);
     } else if ($nilai < 1000) {
-        $temp = terbilang($nilai / 100) . " Ratus" . terbilang($nilai % 100);
+        $temp = terbilang((int)($nilai / 100)) . " Ratus" . terbilang($nilai % 100);
     } else if ($nilai < 2000) {
         $temp = " Seribu" . terbilang($nilai - 1000);
     } else if ($nilai < 1000000) {
-        $temp = terbilang($nilai / 1000) . " Ribu" . terbilang($nilai % 1000);
+        $temp = terbilang((int)($nilai / 1000)) . " Ribu" . terbilang($nilai % 1000);
     } else if ($nilai < 1000000000) {
-        $temp = terbilang($nilai / 1000000) . " Juta" . terbilang($nilai % 1000000);
+        $temp = terbilang((int)($nilai / 1000000)) . " Juta" . terbilang($nilai % 1000000);
     } else if ($nilai < 1000000000000) {
-        $temp = terbilang($nilai / 1000000000) . " Milyar" . terbilang($nilai % 1000000000);
+        $temp = terbilang((int)($nilai / 1000000000)) . " Milyar" . terbilang($nilai % 1000000000);
     } else if ($nilai < 1000000000000000) {
-        $temp = terbilang($nilai / 1000000000000) . " Trilyun" . terbilang($nilai % 1000000000000);
+        $temp = terbilang((int)($nilai / 1000000000000)) . " Trilyun" . terbilang($nilai % 1000000000000);
     }
     
     return trim($temp);
